+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+# v2.1.2 (2022-08-23)
+
+### Fixed
+
+  * Scripts did not log errors to Discord ;
+  * ``update_sub_states`` script crashed because of PCeens without IntraRez access.
+
+
 # v2.1.1 (2022-07-18)
 
 ### Changed